Fixed enemy aiming bug
Enemies were not aiming correctly so I added a bunch of debug stuff to fix it and also the bullets are circles now because circles are easier than rectangles and also I was just doing math wrong but I don't feel like changing the bullets back to rectangles and we might not need to anyways so I don't really care about doing that.

How long can these commit messages be?